ql: modify table to handle unknown number of actions
